% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_trackplots_all.R
\name{generate_trackplots_all}
\alias{generate_trackplots_all}
\title{Generate animated snout trackplots for all CSVs in a folder}
\usage{
generate_trackplots_all(folder, fps = 30, width = 800, height = 600)
}
\arguments{
\item{folder}{Path to the folder containing "Relative CSVs".}

\item{fps}{Frames per second for the animation (default = 30).}

\item{width}{Output GIF width in pixels (default = 800).}

\item{height}{Output GIF height in pixels (default = 600).}
}
\value{
A vector of paths to the generated GIF files.
}
\description{
This function loops through all "Relative CSVs" in a folder and generates
animated trajectory plots (.gif) showing the snout tip movement relative
to the base of the tail. Each animation is saved in a "Trackplots" subfolder.
}
