% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_data.R
\name{visualise_data}
\alias{visualise_data}
\title{Visualise DeepLabCut data (trackplots, heatmaps, and polar histograms)}
\usage{
visualise_data(folder)
}
\arguments{
\item{folder}{Path to the parent folder containing "Filtered CSVs" with subfolders:
"Relative CSVs" and/or "Angle CSVs".}
}
\description{
Creates high-quality visualisations:
- Trackplots of snout trajectories
- 2D kernel density heatmaps
- Polar histograms of angles
}
\details{
All figures include labeled axes and legends, and are saved as 600 dpi TIFFs
inside a "Visualisation" folder.
}
